Split pathfinder into multiple smaller files
