+++ AI/Nullkiller/Pathfinding/Rules/AILayerTransitionRule.h
@@ -15,6 +15,7 @@
 #include "../Actions/BoatActions.h"
 #include "../../../../CCallback.h"
 #include "../../../../lib/mapObjects/MapObjects.h"
+#include "../../../../lib/pathfinder/PathfindingRules.h"
 
 namespace NKAI
 {
