+++ AI/Nullkiller/Pathfinding/Rules/AIMovementAfterDestinationRule.h
@@ -14,6 +14,7 @@
 #include "../../AIGateway.h"
 #include "../../../../CCallback.h"
 #include "../../../../lib/mapObjects/MapObjects.h"
+#include "../../../../lib/pathfinder/PathfindingRules.h"
 
 namespace NKAI
 {
