+++ AI/VCAI/Pathfinding/Rules/AIPreviousNodeRule.h
@@ -14,6 +14,7 @@
 #include "../../VCAI.h"
 #include "../../../../CCallback.h"
 #include "../../../../lib/mapObjects/MapObjects.h"
+#include "../../../../lib/pathfinder/PathfindingRules.h"
 
 namespace AIPathfinding
 {
