+++ client/Client.cpp
@@ -24,12 +24,12 @@
 #include "../CCallback.h"
 #include "../lib/CConfigHandler.h"
 #include "../lib/CGameState.h"
-#include "../lib/CPathfinder.h"
 #include "../lib/CThreadHelper.h"
 #include "../lib/VCMIDirs.h"
 #include "../lib/battle/BattleInfo.h"
 #include "../lib/serializer/BinaryDeserializer.h"
 #include "../lib/mapping/CMapService.h"
+#include "../lib/pathfinder/CGPathNode.h"
 #include "../lib/filesystem/Filesystem.h"
 #include "../lib/registerTypes/RegisterTypes.h"
 #include "../lib/serializer/Connection.h"
