+++ lib/pathfinder/PathfinderUtil.h
@@ -9,10 +9,10 @@
  */
 #pragma once
 
-#include "TerrainHandler.h"
-#include "mapObjects/CGObjectInstance.h"
-#include "mapping/CMapDefines.h"
-#include "CGameState.h"
+#include "../TerrainHandler.h"
+#include "../mapObjects/CGObjectInstance.h"
+#include "../mapping/CMapDefines.h"
+#include "../CGameState.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
